requirements.txt
wtforms
sqlalchemy
flask
flask-sqlalchemy
flask-bcrypt
flask-login
flask-wtf
flask-markdown
email_validator
Pillow
requests
pytest
selenium
lorem-text
